build: Move build related unit test into own job
We shouldn't be asserting on built assets (dist/esm) in our unit tests.

This helps unblock https://github.com/getsentry/sentry-javascript/pull/4616
