+++ .github/workflows/build.yml
@@ -411,7 +411,11 @@ jobs:
         with:
           path: ${{ env.CACHED_BUILD_PATHS }}
           key: ${{ env.BUILD_CACHE_KEY }}
-      - name: Run build tests
+      - name: Run browser build tests
         run: |
           cd packages/browser
+          yarn test:package
+      - name: Run utils build tests
+        run: |
+          cd packages/utils
           yarn test:package
