+++ packages/utils/.gitignore
@@ -1,4 +1,5 @@
 *.js.map
 *.d.ts
 *.js
+!test/package/build.js
 !.eslintrc.js
