+++ packages/utils/package.json
@@ -45,7 +45,8 @@
     "lint:prettier": "prettier --check \"{src,test}/**/*.ts\"",
     "pack": "npm pack",
     "test": "jest",
-    "test:watch": "jest --watch"
+    "test:watch": "jest --watch",
+    "test:package": "node test/package/build.js"
   },
   "volta": {
     "extends": "../../package.json"
